Payloader
Payloader- The script runs an automation payloader, which gives the user an IP address and a port number to listen with and runs automatically. It gets a listener and opens an http server to send the link to the victim. 
The script includes penetration testing tools, such as Msfvenom and Msfconsole.